+++ app/src/main/java/com/tellingus/tellingme/data/model/login/TokenDto.kt
@@ -1,0 +1,7 @@
+package com.tellingus.tellingme.data.model.login
+
+data class TokenDto(
+    val accessToken: String = "",
+    val refreshToken: String = "",
+    val socialId: String = "",
+)
